pyproject.toml
[project]
name = "lngraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.0",
    "langchain-community>=0.3.27",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.2",
    "python-dotenv>=1.1.1",
]
